api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
scikit-learn==1.3.0
joblib==1.3.2
email-validator==2.1.0
pydantic[email]==2.5.0